+++ client/src/pages/Design.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import '../../App.css';
+
+export default function Design() {
+  return (
+    <>
+      <h1 className='design'>DESIGN</h1>
+    </>
+  );
+}
