+++ client/src/pages/Development.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import '../../App.css';
+
+export default function Development() {
+  return (
+    <>
+      <h1 className='development'>DEVELOPMENT</h1>
+    </>
+  );
+}
