+++ client/src/pages/Marketing.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import '../../App.css';
+
+export default function Marketing() {
+  return (
+    <>
+      <h1 className='marketing'>MARKETING</h1>
+    </>
+  );
+}
