--- pages/Design.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import '../../App.css';
-
-export default function Design() {
-  return (
-    <>
-      <h1 className='design'>DESIGN</h1>
-    </>
-  );
-}
