--- pages/Home.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import '../../App.css';
-
-export default function Home() {
-  return (
-    <>
-      <h1 className='home'>GROWGREEN</h1>
-    </>
-  );
-}
